+++ src/main/java/com/barion/dungeons_enhanced/world/structures/DEDesertTomb.java
@@ -28,8 +28,8 @@ public class DEDesertTomb{
     }*/
 
     public static class Capability implements JigsawCapability.IJigsawCapability{
-        public static final Capability Instance = new Capability();
-        public static final Codec<Capability> CODEC = Codec.unit(Instance);
+        public static final Capability INSTANCE = new Capability();
+        public static final Codec<Capability> CODEC = Codec.unit(INSTANCE);
 
         @Override
         public JigsawCapability.JigsawType<?> getType(){return DEJigsawTypes.DesertTomb;}
@@ -42,7 +42,7 @@ public class DEDesertTomb{
         public Piece(StructurePieceSerializationContext serializationContext, CompoundTag nbt) {super(serializationContext, nbt);}
 
         @Override
-        public StructurePieceType getType() {return DEStructures.DesertTomb.getPieceType().get();}
+        public StructurePieceType getType() {return DEStructures.DESERT_TOMB.getPieceType().get();}
         @Override
         public void handleDataMarker(String key, BlockPos pos, ServerLevelAccessor levelAccessor, RandomSource random, BoundingBox box) {}
     }